+++ tests/invalid_nodes.py
@@ -10,3 +10,13 @@ class InvalidUserNode(nodes.InvalidNode):
 
     def key(self):
         return 'user:%s' % self.user
+
+
+class StaleInvalidNode(nodes.InvalidNode):
+    def key(self):
+        return 'test_stale'
+
+
+class NoStaleInvalidNode(nodes.InvalidNode):
+    def key(self):
+        return 'test_np_stale'